Refactor to make modules self-contained
